Remove EventSourcedAggregate related files, attrs and methods
