+++ src/protean/utils/__init__.py
@@ -118,7 +118,6 @@ class DomainObjects(Enum):
     COMMAND_HANDLER = "COMMAND_HANDLER"
     EVENT = "EVENT"
     EVENT_HANDLER = "EVENT_HANDLER"
-    EVENT_SOURCED_AGGREGATE = "EVENT_SOURCED_AGGREGATE"
     EVENT_SOURCED_REPOSITORY = "EVENT_SOURCED_REPOSITORY"
     DOMAIN_SERVICE = "DOMAIN_SERVICE"
     EMAIL = "EMAIL"
