+++ tests/event_store/test_appending_aggregate_events.py
@@ -4,8 +4,7 @@ from uuid import uuid4
 
 import pytest
 
-from protean import BaseAggregate, BaseEvent
-from protean.core.event_sourced_aggregate import apply
+from protean import BaseAggregate, BaseEvent, apply
 from protean.fields import String
 from protean.fields.basic import Identifier
